+++ src/puget/data.clj
@@ -1,7 +1,5 @@
 (ns puget.data
   "Code to handle custom data represented as tagged EDN values."
-  (:require
-    [clojure.data.codec.base64 :as b64])
   (:import
     (java.net URI)
     (java.util Date TimeZone UUID)))
@@ -16,6 +14,8 @@
     serialization. Returns a `TaggedLiteral` record."))
 
 
+; TODO: remove this when CLJ-1424 merges
+; http://dev.clojure.org/jira/browse/CLJ-1424
 (defrecord TaggedLiteral
   [tag form]
 
@@ -33,6 +33,11 @@
     (str \# tag \space (pr-str form))))
 
 
+;; Remove automatic constructor functions.
+(ns-unmap *ns* '->TaggedLiteral)
+(ns-unmap *ns* 'map->TaggedLiteral)
+
+
 (defmethod print-method TaggedLiteral
   [v ^java.io.Writer w]
   (.write w (str v)))
@@ -43,21 +48,13 @@
   suitable for use as a default-data-reader function."
   [tag value]
   {:pre [(symbol? tag)]}
-  (->TaggedLiteral tag value))
+  (TaggedLiteral. tag value))
 
 
 (defn tagged-literal?
   "Returns true if the given value is a tagged-literal form."
   [value]
   (instance? TaggedLiteral value))
-
-
-(defn edn-str
-  "Converts the given value to a tagged EDN string. Falls back to `pr-str` if
-  `v` does not use extended notation."
-  ^String
-  [v]
-  (pr-str (if (satisfies? ExtendedNotation v) (->edn v) v)))
 
 
 
@@ -92,7 +89,7 @@
 
 
 
-;; ## Basic EDN Types
+;; ## Standard EDN Types
 
 (defn- format-utc
   "Produces an ISO-8601 formatted date-time string from the given Date."
@@ -109,30 +106,3 @@
 
 ;; `uuid` tags a universally-unique identifier string.
 (extend-tagged-str UUID 'uuid)
-
-
-;; `puget/bin` tags byte data represented as a base64-encoded string.
-(extend-tagged-value
-  (class (byte-array 0))
-  'puget/bin
-  #(->> % b64/encode (map char) (apply str)))
-
-
-(defn read-bin
-  "Reads a base64-encoded string into a byte array. Suitable as a data-reader
-  for `puget/bin` literals."
-  ^bytes
-  [^String bin]
-  (b64/decode (.getBytes bin)))
-
-
-;; `puget/uri` tags a Universal Resource Identifier string.
-(extend-tagged-str URI 'puget/uri)
-
-
-(defn read-uri
-  "Constructs a URI from a string value. Suitable as a data-reader for
-  `puget/uri` literals."
-  ^URI
-  [^String uri]
-  (URI. uri))